+++ src/main/webapp/instrument.js
@@ -1,4 +1,31 @@
-function init(id) {
+function init(id, loansTable, statusesTable) {
+    $.getJSON('webresources/instruments/' + id + '/loans', function(data) {
+        loansTable.clear();
+        for(i=0; i<data.length; i++) {
+           loansTable.row.add([
+                data[i].musicianName,
+                data[i].outAt==null?"":new Date(data[i].outAt).toLocaleDateString(),
+                data[i].outByUser,
+                data[i].inAt==null?"":new Date(data[i].inAt).toLocaleDateString(),
+                data[i].inByUser
+            ]);
+        }
+        loansTable.draw();       
+    });
+    
+    $.getJSON('webresources/instruments/' + id + '/statuses', function(data) {
+        statusesTable.clear();
+        for(i=0; i<data.length; i++) {
+           statusesTable.row.add([
+                data[i].text,
+                data[i].date==null?"":new Date(data[i].date).toLocaleDateString(),
+                data[i].statusByUser
+            ]);
+        }
+        statusesTable.draw();       
+    });
+    
+    $('#type').empty();
     $.getJSON('webresources/instruments/types', function(data) {
         $.each(data, function(key, value) {   
             $('#type')
@@ -8,6 +35,7 @@ function init(id) {
         });        
     });
     
+    $('#makes').empty();
     $.getJSON('webresources/instruments/makes', function(data) {
         $.each(data, function(key, value) {   
             $('#makes')
@@ -16,6 +44,7 @@ function init(id) {
         });        
     });
 
+    $('#musician').empty();
     $.getJSON('webresources/musicians', function(data) {
         $.each(data, function(key, value) {   
             $('#musician')
@@ -36,15 +65,41 @@ function init(id) {
         $('#productNo').val(data.productNo);
         $('#serialNo').val(data.serialNo);
         $('#description').val(data.description);
-        $('#lentTo').html(data.lentTo);      
+        $('#lentTo').html(data.lentTo);   
+        $('#lastStatus').html(data.status);
+        
+        if(data.lentTo == null || data.lentTo == "") {
+            $("#endLoan").hide();
+            $("#newLoan").show();
+        } else {            
+            $("#newLoan").hide();
+            $("#endLoan").show();
+        }
     });
 }
     
 $(document).ready(function() {
     
     var id = getQueryParam("id");
+
+    var loansTable = $('#loans').DataTable( {
+        "paging": false,
+        "lengthChange": false,
+        "info": false,
+        "searching": false,
+        "order": [[ 1, "desc" ]]
+    });
+
+    var statusesTable = $('#statuses').DataTable( {
+        "paging": false,
+        "lengthChange": false,
+        "info": false,
+        "searching": false,
+        "order": [[ 1, "desc" ]]
+    });
+
     
-    init(id);
+    init(id, loansTable, statusesTable);
     
     getLoggedOnUser(function(data) {
         $("#logout").html(data.email + ": logg ut");
@@ -68,7 +123,7 @@ $(document).ready(function() {
             }),
             contentType: "application/json",
             success: function (data, textStatus, jqXHR) {
-                init(id);
+                init(id, loansTable, statusesTable);
                 $('#editInstrumentModal').modal('hide');
             },
             error: function (jqXHR, textStatus, errorThrown) {
@@ -77,6 +132,26 @@ $(document).ready(function() {
         });    
     });
 
+    $("#newStatusForm").submit(function(event) {
+        // Stop form from submitting normally
+        event.preventDefault();
+        
+        // Save new state
+        $.ajax({
+            type: "POST",
+            url: "webresources/instruments/" + id + "/statuses",
+            data: $("#status").val(),
+            contentType: "text/plain",
+            success: function (data, textStatus, jqXHR) {
+                init(id, loansTable, statusesTable);
+                $('#newStatusModal').modal('hide');
+            },
+            error: function (jqXHR, textStatus, errorThrown) {
+                window.location.href = "error.html";
+            }
+        });    
+    });
+    
     $("#newLoanForm").submit(function(event) {
         // Stop form from submitting normally
         event.preventDefault();
@@ -84,11 +159,11 @@ $(document).ready(function() {
         // Save new loan
         $.ajax({
             type: "POST",
-            url: "webresources/instruments/" + id + "/loan",
+            url: "webresources/instruments/" + id + "/loans",
             data: $("#musician").val(),
             contentType: "text/plain",
             success: function (data, textStatus, jqXHR) {
-                init(id);
+                init(id, loansTable, statusesTable);
                 $('#newLoanModal').modal('hide');
             },
             error: function (jqXHR, textStatus, errorThrown) {
@@ -98,18 +173,19 @@ $(document).ready(function() {
     });
     
     $("#endLoan").click(function(event) {
-        bootbox.confirm("Er du sikker på at du vil avslutte utlån?", function(result) {    
-            $.ajax({
-                url: 'webresources/instruments/' + id + '/loan',
-                type: 'DELETE',
-                success: function(data) {
-                    init(id);
-                },
-                error: function() {
-                    window.location.href = "error.html";
-                }                   
-            });
+        bootbox.confirm("Er du sikker på at du vil avslutte utlån?", function(result) {   
+            if(result) {
+                $.ajax({
+                    url: 'webresources/instruments/' + id + '/loans',
+                    type: 'DELETE',
+                    success: function(data) {
+                        init(id, loansTable, statusesTable);
+                    },
+                    error: function() {
+                        window.location.href = "error.html";
+                    }                   
+                });
+            }
         });                  
     });
-
 });
